+++ src/tractusx_sdk/dataspace/constants.py
@@ -20,6 +20,5 @@
 # SPDX-License-Identifier: Apache-2.0
 #################################################################################
 
-
 DSP_DATASET_KEY:str="dcat:dataset"
 DSP_POLICY_KEY:str="odrl:hasPolicy"
