+++ src/tractusx_sdk/dataspace/controllers/connector/controller_factory.py
@@ -53,7 +53,7 @@ class ControllerFactory:
     SUPPORTED_VERSIONS = []
     for module in listdir(_controllers_base_path):
         module_path = path.join(_controllers_base_path, module)
-        if path.isdir(module_path) and module.startswith("v"):
+        if path.isdir(module_path) and module != "__pycache__" and module != "utils":
             SUPPORTED_VERSIONS.append(module)
 
     @staticmethod
@@ -69,7 +69,7 @@ class ControllerFactory:
         controller class, and returns it, with whatever parameters necessary for its initialization.
 
         :param controller_type: The type of controller to create, as per the AdapterType enum
-        :param connector_version: The version of the Connector (e.g., "v0_9_0")
+        :param connector_version: The version of the Connector (e.g., "jupiter")
         :return: An instance of the specified Adapter subclass
         """
 
@@ -107,7 +107,7 @@ class ControllerFactory:
         """
         Create an asset controller instance, based a specific version.
 
-        :param connector_version: The version of the Connector (i.e: "v0_9_0")
+        :param connector_version: The version of the Connector (i.e: "jupiter")
         :param adapter: The DMA adapter to use for the controller
 
         :return: An instance of the specified Controller subclass
@@ -133,7 +133,7 @@ class ControllerFactory:
         """
         Create a catalog controller instance, based a specific version.
 
-        :param connector_version: The version of the Connector (i.e: "v0_9_0")
+        :param connector_version: The version of the Connector (i.e: "jupiter")
         :param adapter: The DMA adapter to use for the controller
 
         :return: An instance of the specified Controller subclass
@@ -159,7 +159,7 @@ class ControllerFactory:
         """
         Create a contract_agreement controller instance, based a specific version.
 
-        :param connector_version: The version of the Connector (i.e: "v0_9_0")
+        :param connector_version: The version of the Connector (i.e: "jupiter")
         :param adapter: The DMA adapter to use for the controller
 
         :return: An instance of the specified Controller subclass
@@ -185,7 +185,7 @@ class ControllerFactory:
         """
         Create a contract_definition controller instance, based a specific version.
 
-        :param connector_version: The version of the Connector (i.e: "v0_9_0")
+        :param connector_version: The version of the Connector (i.e: "jupiter")
         :param adapter: The DMA adapter to use for the controller
 
         :return: An instance of the specified Controller subclass
@@ -211,7 +211,7 @@ class ControllerFactory:
         """
         Create a contract_negotiation controller instance, based a specific version.
 
-        :param connector_version: The version of the Connector (i.e: "v0_9_0")
+        :param connector_version: The version of the Connector (i.e: "jupiter")
         :param adapter: The DMA adapter to use for the controller
 
         :return: An instance of the specified Controller subclass
@@ -237,7 +237,7 @@ class ControllerFactory:
         """
         Create an EDR controller instance, based a specific version.
 
-        :param connector_version: The version of the Connector (i.e: "v0_9_0")
+        :param connector_version: The version of the Connector (i.e: "jupiter")
         :param adapter: The DMA adapter to use for the controller
 
         :return: An instance of the specified Controller subclass
@@ -263,7 +263,7 @@ class ControllerFactory:
         """
         Create a policy controller instance, based a specific version.
 
-        :param connector_version: The version of the Connector (i.e: "v0_9_0")
+        :param connector_version: The version of the Connector (i.e: "jupiter")
         :param adapter: The DMA adapter to use for the controller
 
         :return: An instance of the specified Controller subclass
@@ -289,7 +289,7 @@ class ControllerFactory:
         """
         Create a transfer_process controller instance, based a specific version.
 
-        :param connector_version: The version of the Connector (i.e: "v0_9_0")
+        :param connector_version: The version of the Connector (i.e: "jupiter")
         :param adapter: The DMA adapter to use for the controller
 
         :return: An instance of the specified Controller subclass
@@ -315,7 +315,7 @@ class ControllerFactory:
         """
         Create all DMA controllers for a specific connector version.
 
-        :param connector_version: The version of the Connector (i.e: "v0_9_0")
+        :param connector_version: The version of the Connector (i.e: "jupiter")
         :param adapter: The DMA adapter to use for the controller
         :param kwargs: Additional parameters to pass to the controller builder
         :return: A dictionary of controller instances, keyed by controller type
