+++ src/tractusx_sdk/dataspace/controllers/connector/jupiter/policy_controller.py
@@ -21,24 +21,22 @@
 #################################################################################
 
 from .dma_controller import DmaController
-from ..mixins import CreateControllerMixin, GetAllControllerMixin, DeleteControllerMixin
-from ....models.connector.v0_9_0 import ContractNegotiationModel
+from tractusx_sdk.dataspace.controllers.connector.utils.mixins import CrudDmaController
+from ....models.connector.jupiter import PolicyModel
 
 
-class EdrController(CreateControllerMixin, GetAllControllerMixin, DeleteControllerMixin, DmaController):
+class PolicyController(CrudDmaController, DmaController):
     """
-    Concrete implementation of the EdrController for the Connector v0.9.0 Data Management API.
+    Concrete implementation of the PolicyController for the Connector v0.9.0 Data Management API.
 
-    This class overrides the create method in order to ensure the correct class types are used, instead of the generic ones.
+    This class overrides the create and update methods in order to ensure the correct class types are used,
+    instead of the generic ones.
     """
 
-    endpoint_url = "/v3/edrs"
-    
-    def get_data_address(self, oid: str, **kwargs):
-        return self.adapter.get(url=f"{self.endpoint_url}/{oid}/dataaddress", **kwargs)
+    endpoint_url = "/v3/policydefinitions"
 
-    def refresh(self, oid: str, **kwargs):
-        return self.adapter.post(url=f"{self.endpoint_url}/{oid}/refresh", **kwargs)
-
-    def create(self, obj: ContractNegotiationModel, **kwargs):
+    def create(self, obj: PolicyModel, **kwargs):
         return super().create(obj, **kwargs)
+
+    def update(self, obj: PolicyModel, **kwargs):
+        return super().update(obj, **kwargs)
