+++ src/tractusx_sdk/dataspace/controllers/connector/utils/decorators/controller_method.py
@@ -21,7 +21,7 @@
 #################################################################################
 
 
-from ...controller import Controller
+from tractusx_sdk.dataspace.controllers.controller import Controller
 
 
 def controller_method(func):
