+++ src/tractusx_sdk/dataspace/models/connector/jupiter/contract_definition_model.py
