+++ src/tractusx_sdk/dataspace/models/connector/jupiter/queryspec_model.py
