+++ tests/dataspace/models/connector/test_model_factory_queryspec.py
@@ -28,7 +28,7 @@ from tractusx_sdk.dataspace.models.connector.model_factory import ModelFactory
 
 class TestModelFactoryQuerySpec(unittest.TestCase):
     def setUp(self):
-        self.connector_version = "v0_9_0"
+        self.connector_version = "jupiter"
         self.context = {"key": "value"}
         self.offset = 5
         self.limit = 20
