+++ tests/dataspace/models/connector/test_model_factory_transfer_process.py
@@ -28,7 +28,7 @@ from tractusx_sdk.dataspace.models.connector.model_factory import ModelFactory
 
 class TestModelFactoryTransferProcess(unittest.TestCase):
     def setUp(self):
-        self.connector_version = "v0_9_0"
+        self.connector_version = "jupiter"
         self.counter_party_address = "https://counterparty.com"
         self.transfer_type = "dataspace-protocol-http"
         self.contract_id = "contract-id"
